Ansible update fixes (#6)
* Update repo URL

The use of the git:// protocol is now deprecated and no longer used, therefore it has been changed to https as this more secure and supported.

* Restrict ansible-lint version

* Setting ansible_python_interpreter

As of the update, ansible does not default to the platform-python. This causes errors with the SELinux commands as the module they require is in python3.6, not 3.8, which ansible defaults to. Using platform Linux as it is is python3.6 solves this issue.

* Updating Create influxdb telegraf datasource task.

This task fails on the new version of ansible on the epel repo as the use of url now refers to the grafana_url, not the data source as it did previously. Making the url explicit with ds_url solves this problem.

Co-authored-by: = <=>
Co-authored-by: Matt Williams <matt@milliams.com>