Ansible update fixes (#9)
* Setting ansible_python_interpreter

As of the update, ansible does not default to the platform-python. This causes errors with the SELinux commands as the module they require is in python3.6, not 3.8, which ansible defaults to. Using platform Linux as it is is python3.6 solves this issue.

* Updating Create influxdb telegraf datasource task.

This task fails on the new version of ansible on the epel repo as the use of url now refers to the grafana_url, not the data source as it did previously. Making the url explicit with ds_url solves this problem.

Co-authored-by: = <=>